Refactored updaters to have a common base
