+++ FireStarterProject/FireStarter/src/main/java/de/belu/firestarter/observer/ForeGroundService.java
@@ -25,7 +25,7 @@ import de.belu.firestarter.R;
 import de.belu.firestarter.gui.MainActivity;
 import de.belu.firestarter.tools.AppStarter;
 import de.belu.firestarter.tools.SettingsProvider;
-import de.belu.firestarter.tools.Updater;
+import de.belu.firestarter.tools.FireStarterUpdater;
 
 
 /**
@@ -247,8 +247,8 @@ public class ForeGroundService extends Service
                 try
                 {
                     Log("Start scheduled update-check:");
-                    Updater updater = new Updater();
-                    updater.checkForUpdate(true);
+                    FireStarterUpdater fireStarterUpdater = new FireStarterUpdater();
+                    fireStarterUpdater.checkForUpdate(true);
                 }
                 catch(Exception e)
                 {
