+++ FireStarterProject/FireStarter/src/main/java/de/belu/firestarter/tests/KodiTests.java
@@ -2,15 +2,15 @@ package de.belu.firestarter.tests;
 
 import android.test.InstrumentationTestCase;
 
-import de.belu.firestarter.tools.UpdaterKodi;
+import de.belu.firestarter.tools.KodiUpdater;
 
 /**
  * Created by attila.szasz on 23-Oct-15.
  */
 public class KodiTests extends InstrumentationTestCase {
     public void testCheckForUpdate() throws Exception {
-        UpdaterKodi mUpdaterKodi = new UpdaterKodi();
-        mUpdaterKodi.checkForUpdate(true);
-        assertNotNull("Latest version should not be null", mUpdaterKodi.LATEST_VERSION);
+        KodiUpdater mKodiUpdater = new KodiUpdater();
+        mKodiUpdater.checkForUpdate(true);
+        assertNotNull("Latest version should not be null", mKodiUpdater.LATEST_VERSION);
     }
 }
